+++ packages/core/src/node/messaging/messaging-contribution.ts
@@ -30,8 +30,8 @@ import { BackendApplicationContribution } from '../backend-application';
 import { MessagingService, WebSocketChannelConnection } from './messaging-service';
 import { ConsoleLogger } from './logger';
 import { ConnectionContainerModule } from './connection-container-module';
-
 import Route = require('route-parser');
+import { WsRequestValidator } from '../ws-request-validators';
 
 export const MessagingContainer = Symbol('MessagingContainer');
 
@@ -46,6 +46,9 @@ export class MessagingContribution implements BackendApplicationContribution, Me
 
     @inject(ContributionProvider) @named(MessagingService.Contribution)
     protected readonly contributions: ContributionProvider<MessagingService.Contribution>;
+
+    @inject(WsRequestValidator)
+    protected readonly wsRequestValidator: WsRequestValidator;
 
     protected webSocketServer: ws.Server | undefined;
     protected readonly wsHandlers = new MessagingContribution.ConnectionHandlers<ws>();
@@ -115,8 +118,20 @@ export class MessagingContribution implements BackendApplicationContribution, Me
      * Route HTTP upgrade requests to the WebSocket server.
      */
     protected handleHttpUpgrade(request: http.IncomingMessage, socket: net.Socket, head: Buffer): void {
-        this.webSocketServer!.handleUpgrade(request, socket, head, client => {
-            this.webSocketServer!.emit('connection', client, request);
+        this.wsRequestValidator.allowWsUpgrade(request).then(allowed => {
+            if (allowed) {
+                this.webSocketServer!.handleUpgrade(request, socket, head, client => {
+                    this.webSocketServer!.emit('connection', client, request);
+                });
+            } else {
+                console.error(`refused a websocket connection: ${request.connection.remoteAddress}`);
+                socket.write('HTTP/1.1 403 Forbidden\n\n');
+                socket.destroy();
+            }
+        }, error => {
+            console.error(error);
+            socket.write('HTTP/1.1 500 Internal Error\n\n');
+            socket.destroy();
         });
     }