+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Change Log
 
+### v1.8.1 - 08/12/2020
+
+- [core] added `THEIA_HOSTS` environment variable (browser applications only).
+  - Used to filter incoming WebSocket connections: if `Origin` header does not match the list of hosts it will be refused.
+  - Value is a comma-separated list of domain names including the port if not `80` nor `443`.
+  - Example: `app.some.domain.com,app.other.domain:12345`.
+
+<a name="breaking_changes_1.8.1">[Breaking Changes:](#breaking_changes_1.8.1)</a>
+
+- [core] Deprecated `ElectronMessagingContribution`, token validation is now done in `ElectronTokenValidator` as a `WsRequestValidatorContribution`.
+- [mini-browser] New unique endpoint.
+  - `{{uuid}}.mini-browser.{{hostname}}` by default.
+  - Can be configured via `THEIA_MINI_BROWSER_HOST_PATTERN` environment variable.
+  - Clients must setup this new hostname in their DNS resolvers.
+
 ## v1.8.0 - 26/11/2020
 
 - [api-tests] fixed issue with `saveable` test suite [#8736](https://github.com/eclipse-theia/theia/pull/8736)